+++ bantam/lib/datasource/index.js
@@ -4,10 +4,14 @@ var _ = require('underscore');
 var log = require(__dirname + '/../log');
 var BearerAuthStrategy = require(__dirname + '/../auth/bearer');
 
+/**
+ * Represents a Datasource.
+ * @constructor
+ */
 var Datasource = function (page, datasource, options, callback) {
 
   this.log = log.get().child({module: 'datasource'});
-  this.log.info('Datasource logging started (' + datasource + ').')
+  //this.log.info('Datasource logging started (' + datasource + ').')
 
   this.page = page;
   this.name = datasource;
@@ -41,6 +45,19 @@ var Datasource = function (page, datasource, options, callback) {
 
 };
 
+/**
+ * Callback for loading a datasource schema.
+ *
+ * @callback loadDatasourceCallback
+ * @param {Error} err - An error occurred whilst trying to load the datasource schema.
+ * @param {JSON} result - the datasource schema.
+ */
+
+/**
+ *  Reads a datasource schema from the filesystem
+ *  @param {loadDatasourceCallback} done - the callback that handles the response
+ *  @public
+ */
 Datasource.prototype.loadDatasource = function(done) {
 
   var filepath = (this.options.datasourcePath || '') + '/' + this.name + '.json';
@@ -85,6 +102,12 @@ Datasource.prototype.setAuthStrategy = function() {
   return new BearerAuthStrategy(this.schema.datasource.auth);
 };
 
+/**
+ *  Constructs the datasource endpoint using properties defined in the schema
+ *  @param {JSON} schema - the callback that handles the response
+ *  @param done - the callback that handles the response
+ *  @public
+ */
 Datasource.prototype.buildEndpoint = function(schema, done) {
 
   if (schema.datasource.source.type === 'static') return;
@@ -102,6 +125,12 @@ Datasource.prototype.buildEndpoint = function(schema, done) {
   done();
 };
 
+/**
+ *  Adds querystring parameters to the datasource endpoint using properties defined in the schema
+ *  @param {JSON} schema - the datasource schema
+ *  @param {String} uri - the original datasource endpoint
+ *  @public
+ */
 Datasource.prototype.processDatasourceParameters = function (schema, uri) {
 
   var query = '?';
@@ -136,13 +165,8 @@ Datasource.prototype.processDatasourceParameters = function (schema, uri) {
 Datasource.prototype.processRequest = function (datasource, req) {
 
   var self = this;
-
+  var originalFilter = _.clone(this.schema.datasource.filter);
   var query = url.parse(req.url, true).query;
-
-  // handle the json flag
-  // if (query.hasOwnProperty('json')) {
-  //   delete query.json;
-  // }
 
   // handle the cache flag
   if (query.hasOwnProperty('cache') && query.cache === 'false') {
@@ -155,7 +179,7 @@ Datasource.prototype.processRequest = function (datasource, req) {
 
   // if the current datasource matches the page name
   // add some params from the query string or request params
-  if (this.page.name && datasource.indexOf(this.page.name) >= 0) {
+  if ((this.page.name && datasource.indexOf(this.page.name) >= 0) || this.page.passFilters === 'true') {
 
     // handle pagination param
     this.schema.datasource.page = query.page || req.params.page || 1;